agent_from_scratch/backend/requirements.txt
google-genai
openai
python-dotenv
fastapi
uvicorn[standard]
